Replaced IComponents with AComponents
Replaced the IComponents interface with the AComponents abstract class to reduce lines of code used within component classes.
